rpa_v2/requirements.txt
selenium~=4.34.2
webdriver-manager~=4.0.2
requests~=2.32.4
python-dotenv~=1.1.1
pandas~=2.3.1
openpyxl~=3.1.5